Rename useFetchUser to useUser
