+++ packages/volto/src/hooks/index.js
@@ -1,3 +1,3 @@
 export { default as useClipboard } from '@plone/volto/hooks/clipboard/useClipboard';
 export { useClient } from '@plone/volto/hooks/client/useClient';
-export { default as useFetchUser } from '@plone/volto/hooks/user/useFetchUser';
+export { default as useUser } from '@plone/volto/hooks/user/useUser';
